mds: send unlinked dentry's inode attributes in the reply trace for unlink
When an inode has multiple dentries (i.e. is hardlinked) then it won't
necessarily get dropped from the cache after an unlink. The MDS doesn't
currently send the updated target inode info in the UNLINK reply traces
however. Have it do that in all unlink replies -- the inode is almost
certainly already in-core so there's little cost to doing so.

Signed-off-by: Jeff Layton <jlayton@redhat.com>